Fix focus loss on modal close
