+++ src/components/Dialog/DialogDisclosure.js
@@ -1,3 +1,5 @@
+/* eslint-disable react/display-name */
+import React from 'react'
 import styled from 'styled-components'
 import { DialogDisclosure as ReakitDialogDisclosure } from 'reakit/Dialog'
 
@@ -8,8 +10,8 @@ const Disclosure = styled.div`
   outline: none;
 `
 
-const DialogDisclosure = props => (
-  <ReakitDialogDisclosure as={Disclosure} {...props} />
-)
+const DialogDisclosure = React.forwardRef((props, ref) => (
+  <ReakitDialogDisclosure as={Disclosure} ref={ref} {...props} />
+))
 
 export default DialogDisclosure
